requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
Django==4.2.4
django-crispy-forms==2.0
django-extensions==3.2.3
django-tables2==2.6.0
django-view-breadcrumbs==2.4.1
Faker==19.3.0
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
